AUTH-1169 - Add integration tests for the AuthorizeAccessTokenHandler
